fix: prevent sdk methods from being removed
useFetch removes methods from objects returned in setup(). Moved useFetch and content to a separate component.
