+++ layouts/default.vue
@@ -1,6 +1,5 @@
 <template>
   <space-provider
-    v-if="content"
     :config="$config.nacelle"
     :space="initialSpace"
     :locale="$config.nacelle.locale"
@@ -9,14 +8,7 @@
   >
     <event-provider>
       <cart-provider>
-        <search-provider :searchData="content.searchData">
-          <site-header :content="content.header" />
-          <cart :content="content.cart" />
-          <nuxt />
-          <site-newsletter :content="content.newsletter" />
-          <site-footer :content="content.footer" />
-          <site-nav :content="content.header" />
-        </search-provider>
+        <site-container />
       </cart-provider>
     </event-provider>
   </space-provider>
@@ -28,67 +20,28 @@ import {
   provide,
   ref,
   useContext,
-  useFetch,
-  watch
+  watch,
 } from "@nuxtjs/composition-api";
 
-import {
-  SpaceProvider,
-  EventProvider,
-  CartProvider,
-  SearchProvider
-} from "@nacelle/vue";
-import SiteHeader from "~/components/header/Header.vue";
-import SiteNewsletter from "~/components/newsletter/Newsletter.vue";
-import SiteFooter from "~/components/footer/Footer.vue";
-import SiteNav from "~/components/nav/Nav.vue";
-import Cart from "~/components/cart/Cart.vue";
+import { SpaceProvider, EventProvider, CartProvider } from "@nacelle/vue";
+import SiteContainer from "~/components/core/SiteContainer.vue";
 
 export default {
   components: {
     SpaceProvider,
     EventProvider,
     CartProvider,
-    SearchProvider,
-    SiteHeader,
-    SiteNewsletter,
-    SiteFooter,
-    SiteNav,
-    Cart
+    SiteContainer,
   },
   setup() {
-    const content = ref(null);
     const cartOpen = ref(false);
     const navOpen = ref(false);
     const initialSpace = inject("initialSpace");
     const nacelleSdk = inject("nacelleSdk");
     const { route } = useContext();
 
-    const setCartOpen = val => (cartOpen.value = val);
-    const setNavOpen = val => (navOpen.value = val);
-
-    useFetch(async () => {
-      const [header, footer, newsletter, cart, searchData] = await Promise.all([
-        nacelleSdk.data.content({
-          handle: "component-header",
-          type: "componentHeader"
-        }),
-        nacelleSdk.data.content({
-          handle: "component-footer",
-          type: "componentFooter"
-        }),
-        nacelleSdk.data.content({
-          handle: "component-newsletter",
-          type: "componentNewsletter"
-        }),
-        nacelleSdk.data.content({
-          handle: "component-cart",
-          type: "componentCart"
-        }),
-        nacelleSdk.data.allProducts()
-      ]);
-      content.value = { header, footer, newsletter, cart, searchData };
-    });
+    const setCartOpen = (val) => (cartOpen.value = val);
+    const setNavOpen = (val) => (navOpen.value = val);
 
     watch(route, () => {
       cartOpen.value = false;
@@ -100,8 +53,8 @@ export default {
     provide("setCartOpen", setCartOpen);
     provide("setNavOpen", setNavOpen);
 
-    return { initialSpace, content, nacelleSdk };
-  }
+    return { initialSpace, nacelleSdk };
+  },
 };
 </script>
